Change percent to 256

+++ contracts/V1EMPRegistry.sol
@@ -17,7 +17,7 @@ contract V1EMPRegistry is
 	address internal _v1EMPStrategyDeployer;
 	address internal _v1EMPStrategyUtility;
 
-	uint16 public constant override ONE_HUNDRED_PERCENT = 1e4;
+	uint256 public constant override ONE_HUNDRED_PERCENT = 1e4;
 
 	uint256 public override v1EMPIdTracker;
 	uint256 public override v1EMPStrategyIdTracker;
